+++ custom_components/open3e/definitions/subfeatures/hysteresis.py
