+++ custom_components/open3e/definitions/subfeatures/smart_grid_temperature_offsets.py
