+++ custom_components/open3e/definitions/subfeatures/temperature_cooling.py
