+++ custom_components/open3e/water_heater.py
@@ -13,7 +13,7 @@ from homeassistant.util.json import json_loads
 from .const import VIESSMANN_TEMP_DHW_MIN, \
     VIESSMANN_TEMP_DHW_MAX
 from .coordinator import Open3eDataUpdateCoordinator
-from .definitions.dmw_mode import DmwMode
+from custom_components.open3e.definitions.subfeatures.dmw_mode import DmwMode
 from .definitions.water_heater import WATER_HEATER, Open3eWaterHeaterEntityDescription
 from .entity import Open3eEntity
 from .ha_data import Open3eDataConfigEntry
